Audulus 4.0 Module Library
Initial module library for Audulus 4.0. I added a folder for me to keep unfinished modules. Ideally there are so many more I would love to include, but I have to switch gears for now to documentation and circle back to module creation when that is done.

To do after this initial commit:
- Double check tags, spelling in descriptions
- Add special "-Submodule" tag to hide submodules
- Zero out coordinates