Changed Core Module Tags
I changed these tags to align with the conventions the rest of the modules were using - the LFO tag, for instance, was only used by this one Core LFO, the rest are filed under Modulation. I also got rid of the Core tag just because it might be confusing making it seem like these modules are the "core" of the library when they're not really.