Moved Jersmi submodules and renamed some modules
Things like the Panner (circler) are really called Panning Circler in the module, just aligning that with what the module says it is. Got rid of the underscore in Multisliders in favor of x, and moved Jersmi's submodules into his own folder in WIP so we can re-include them when we have a UI solution for submodules.