Moved buggy example Mundone to WIP folder
I moved Mundone, an example patch that would not start on startup, into a different folder. There is some bug in how it's created, but I didn't want to spend too long trying to fix it. For now, I just moved it to WIP and will come back to it later to try to see what's wrong with it.